formating the codebase for easier readability and small code files
