+++ lib/at_sample_app/controllers/contact_controller.ex
@@ -1,0 +1,8 @@
+defmodule AtSampleApp.ContactController do
+  @moduledoc false
+  use AtSampleAppWeb, :controller
+
+  def contact(conn) do
+    render(conn, "contact.html")
+  end
+end
