+++ lib/at_sample_app/router.ex
@@ -1,29 +1,8 @@
 defmodule AtSampleApp.Router do
-  use Plug.Router
-
-  @template_dir "lib/at_sample_app/templates"
-
-  plug(Plug.Parsers,
-    parsers: [:json],
-    pass: ["text/*"],
-    json_decoder: Jason
-  )
-
-  plug(:match)
-  plug(:dispatch)
+  use AtSampleAppWeb, :router
 
   get "/contact" do
-    render(conn, "contact.html")
-  end
-
-  defp render(%{status: status} = conn, template, assigns \\ []) do
-    body =
-      @template_dir
-      |> Path.join(template)
-      |> String.replace_suffix(".html", ".html.eex")
-      |> EEx.eval_file(assigns)
-
-    send_resp(conn, status || 200, body)
+    AtSampleApp.ContactController.contact(conn)
   end
 
   match _ do
